+++ score-service/src/test/java/org/oagi/score/service/bie/BieUpliftingServiceTest.java
@@ -60,7 +60,7 @@ public class BieUpliftingServiceTest extends AbstractServiceTest {
 
         AnalysisBieUpliftingRequest request = new AnalysisBieUpliftingRequest();
         request.setRequester(requester);
-        request.setTopLevelAsbiepId(topLevelAsbiepRecord.getTopLevelAsbiepId().toBigInteger());
+        request.setTopLevelAsbiepId(topLevelAsbiepRecord.getTopLevelAsbiepId());
         request.setTargetReleaseId(topLevelAsbiepRecord.getReleaseId());
 
         assertThrows(IllegalArgumentException.class, () -> {
@@ -79,7 +79,7 @@ public class BieUpliftingServiceTest extends AbstractServiceTest {
                         .withUserId(topLevelAsbiepRecord.getOwnerUserId())
                 )
                 .getUser());
-        request.setTopLevelAsbiepId(topLevelAsbiepRecord.getTopLevelAsbiepId().toBigInteger());
+        request.setTopLevelAsbiepId(topLevelAsbiepRecord.getTopLevelAsbiepId());
         request.setTargetReleaseId(scoreRepositoryFactory.createReleaseReadRepository()
                 .getRelease(new GetReleaseRequest(request.getRequester())
                         .withReleaseNum("10.7")
